% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation_helper_functions.R
\name{add_country_groups}
\alias{add_country_groups}
\title{Consistently Group Countries}
\usage{
add_country_groups(grouping = NULL)
}
\arguments{
\item{grouping}{Currently \code{NULL}, makes sense only when alternative groupings will be available.}
}
\value{
A named vector with color codes.
}
\description{
Create consistent grouping of countries. When too many values are difficult to
follow on a chart, faceting by country group can be one solution.
}
\examples{
{
add_country_groups()
}
}
\seealso{
Other visualisation functions: 
\code{\link{create_discrete_observatory_palette}()}
}
\concept{visualisation functions}
