% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_keywords.R
\name{add_keywords}
\alias{add_keywords}
\title{Add Keywords}
\usage{
add_keywords(description_table, keywords, description = NULL)
}
\arguments{
\item{keywords}{A list of at least four keywords to place to indicator in the observatory.}

\item{description}{Currently only default \code{NULL}, it will be a vector of human-written descriptions.}

\item{metadata}{A metadata table from \code{\link{get_eurostat_indicator}}.}
}
\value{
A data frame of indicator codes, descriptions, original source codes,
and at least keywords to place them in the observatory.
}
\description{
Keywords help place the indicator in the long-form documentation ("pillar" or "chapter")
of the observatory.
}
\details{
1st keyword: which observatory, for example \code{"music"}
2nd keyword: which pillar, for example \code{"economy"}
3rd keyword: top level division of the pillar.
4th keyword: 2nd level division of the pillar.

Further keywords are added to \code{further_keywords} as a concantenated list.
}
\concept{metadata functions}
