% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate.R
\name{add_new_periods}
\alias{add_new_periods}
\title{Add New Periods}
\usage{
add_new_periods(indic, years = NULL, days = NULL)
}
\arguments{
\item{indic}{A temporary indicator table created by an approximation function.}

\item{years}{The number of years to add to the indicator's data frame.  Positive values add after the last
observed time, negative values add before the first observed time.}

\item{days}{The number of years to add to the indicator's data frame.  Positive values add after the last
observed time, negative values add before the first observed time.}
}
\value{
A new data frame with the new observation times added with missing values, labelled as
\code{estimate='missing'} and \code{method='missing'}.
}
\description{
Create a time series object from tmp in approximation functions.
}
\keyword{internal}
