% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing_labels.R
\name{check_missing_labels}
\alias{check_missing_labels}
\title{Check missing labels in a tidy indicator table.}
\usage{
check_missing_labels(check_indicator, only_actual = TRUE)
}
\arguments{
\item{check_indicator}{A tidy indicator from \code{\link{tidy_indicator}}.}

\item{only_actual}{A boolean, if \code{TRUE}, then the total number of observations
must be equal to the total number of \code{actual} and \code{missing observations}.}
}
\value{
A boolean if all assertions are met.
}
\description{
Check if all missing observations are correctly labelled in the \code{method}
and \code{estimate} variable columns.
}
\keyword{internal}
