% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns_without_nan.R
\name{columns_with_nan}
\alias{columns_with_nan}
\title{Function to get the names of the columns with NAN values}
\usage{
columns_with_nan(data)
}
\arguments{
\item{data}{The DataFrame's rows represent observations and the columns
represent variables}
}
\value{
It returns a vector of columns with NAN
}
\description{
This function identifies and returns the names of the columns in a DataFrame
that contain missing values (NaN). It is particularly useful for missing
data imputation and preliminary analysis, allowing for quick identification
of columns that need to be handled due to the presence of missing values
}
\examples{

data("airquality")
columns_with_nan(airquality)

}
\keyword{data}
\keyword{manipulation}
\keyword{missing}
\keyword{values}
