% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_CI.R
\name{compute_CI}
\alias{compute_CI}
\title{Calculation of Condition Indices}
\usage{
compute_CI(matrix)
}
\arguments{
\item{matrix}{a matrix of data where rows = observations and columns =
variables}
}
\value{
It returns the condition index of the matrix
}
\description{
Diagnosis of collinearity in matrix \code{X}
}
\details{
Collinearities can inflate the variance of the estimated regression
coefficients and numerical stability. The condition indices are calculated
by the eigenvalues of the crossproduct matrix of the scaled but uncentered
explanatory variables. Indices > 30 may indicate collinearity
}
\examples{

data("Education")
compute_CI(Education)

}
\references{
Belsley, D. , Kuh, E. and Welsch, R. E. (1979), Regression
Diagnostics: Identifying Influential Data and Sources of Collinearity, John
Wiley (New York)
}
\keyword{ConditionIndex}
\keyword{collinearity}
