% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_database.R
\name{create_database}
\alias{create_database}
\title{Save Eurostat Indicators into a Database}
\usage{
create_database(
  indicator_tables,
  metadata_tables,
  labelling_table,
  description_table,
  db_path = tempdir()
)
}
\arguments{
\item{indicator_tables}{Indicators created by \code{\link{get_eurostat_indicator}} or \code{\link{impute_indicators}}.}

\item{metadata_tables}{Meatadata tables created by \code{\link{get_eurostat_indicator}} or \code{\link{update_metadata}}.}

\item{labelling_table}{Labelling table created by \code{\link{get_eurostat_indicator}}}

\item{description_table}{Description and keywords created for the indicators
by \code{\link{add_keywords}}.}

\item{db_path}{A path to save the database. Defaults to \code{tempdir()}}
}
\value{
An Sqlite database with three tables: indicator, metadata and labelling.
}
\description{
Download and process Eurostat indicators, create their labelling table and descriptive
metadata about the indicators. Save the two metadata tables into metadata tables, and
the tidy indicators themselves into an indicator table.
}
\examples{
\dontrun{
## See vignette
)
}
}
\concept{database functions}
