% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_discrete_observatory_palette.R,
%   R/visualisation_helper_functions.R
\name{create_discrete_observatory_palette}
\alias{create_discrete_observatory_palette}
\title{Create Observatory Color Palette}
\usage{
create_discrete_observatory_palette(palette_name = "european_countries")

create_discrete_observatory_palette(palette_name = "european_countries")
}
\arguments{
\item{palette_name}{Currently only one palette is defined, \code{"european_countries"} which
will create a consistent color palette for European countries.}
}
\value{
A named vector with color codes.

A named vector with color codes.
}
\description{
Create consistent palettes for visualizing data.

Create consistent palettes for visualizing data. Additional palette defintions
are possible, but it is important that our long-form documentation looks consistent.
}
\examples{
{
create_discrete_observatory_palette ("european_countries" )
}
{
create_discrete_observatory_palette ("european_countries" )
}
}
\seealso{
Other visualisation functions: 
\code{\link{add_country_groups}()}

Other visualisation functions: 
\code{\link{add_country_groups}()}
}
\concept{visualisation functions}
