% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_eurostat_database.R
\name{create_eurostat_database}
\alias{create_eurostat_database}
\title{Map Eurostat indicators into a database}
\usage{
create_eurostat_database(ids, path)
}
\arguments{
\item{ids}{Identifiers of Eurostat statistical products.}

\item{path}{A path to save the database.}
}
\value{
An Sqlite database with two tables: metadata and labelling.
}
\description{
Map Eurostat indicators into a database
}
\examples{
\dontrun{
map_indicator_eurostat <- function (
 ids = c("ISOC_R_BLT12_I", "isoc_cicce_use", "teicp090"),
 path = 'not_included/example.db' )
}
}
