% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_database.R
\name{create_indicator_table}
\alias{create_indicator_table}
\title{Create the indicator table in the database}
\usage{
create_indicator_table(con)
}
\description{
Create the indicator table in the database
}
\keyword{internal}
