% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_database.R
\name{create_labelling_table}
\alias{create_labelling_table}
\title{Initialize the labelling table}
\usage{
create_labelling_table(con)
}
\arguments{
\item{con}{A connection}
}
\description{
Initialize the labelling table
}
\keyword{internal}
