% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_database.R
\name{create_metadata_table}
\alias{create_metadata_table}
\title{Initialize the metadata table}
\usage{
create_metadata_table(con)
}
\arguments{
\item{con}{A connection}
}
\description{
Initialize the metadata table
}
\keyword{internal}
