% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate.R
\name{create_time_series}
\alias{create_time_series}
\title{Create Time Series Object}
\usage{
create_time_series(tmp)
}
\arguments{
\item{tmp}{A temporary indicator table created by an approximation function.}
}
\value{
\code{TRUE} if the test is met, otherwise and error message.
}
\description{
Create a time series object from tmp in approximation functions.
}
\keyword{internal}
