% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Education.R
\docType{data}
\name{Education}
\alias{Education}
\title{Education}
\format{
A data frame with 20 observations on the following 8 variables:
\describe{
Participation in the school system of 4-5 year olds

People with at least a high school diploma (25-64 years old)

Graduates and other tertiary degrees (30-34 years)

Participation in continuing education

People earning a STEM tertiary degree in the year Cultural

Participation outside the home

Reading of books and newspapers

Library use}
}
\source{
https://www.istat.it/en/well-being-and-sustainability/the-measurement-of-well-being/indicators
}
\usage{
Education
}
\description{
This dataset contains educational and cultural data related to Italian
regions. Each row represents a region, and there are 20 Italian regions. The
eight variables included in the dataset provide information on various
aspects of the educational system and cultural activities in the regions in
2020.
}
\keyword{datasets}
