% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_table.R
\name{export_table}
\alias{export_table}
\title{Export a table}
\usage{
export_table(table, db, savefile)
}
\arguments{
\item{table}{A table name.}

\item{db}{A path to the database to import the table from.}

\item{savefile}{A path to save the exported table.}
}
\description{
Export a table
}
