% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing_from_long_form.R
\name{fill_missing_from_long_form}
\alias{fill_missing_from_long_form}
\title{Fill missing indicator values}
\usage{
fill_missing_from_long_form(indic_to_fill)
}
\arguments{
\item{indic_to_fill}{A long form Eurostat indicator received by \code{\link{get_eurostat_indicator}}
that may contain implicitly missing values.}
}
\value{
A tibble explicitly showing missing observations.
}
\description{
Missing values are filled when the original long form data frame
hides the fact that certain observations are not present.
}
\keyword{internal}
