% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate.R
\name{format_return_value}
\alias{format_return_value}
\title{Format returned data table}
\usage{
format_return_value(long_form_approx, type_approx = "locf")
}
\arguments{
\item{One}{of \code{nocb}, \code{locf}, \code{forecast}, \code{backcast}}
}
\description{
Format returned data table
}
\keyword{internal}
