% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eurostat_indicator.R
\name{get_eurostat_indicator}
\alias{get_eurostat_indicator}
\title{Get Tidy Indicator(s) From Eurostat}
\usage{
get_eurostat_indicator(id, eurostat_toc = NULL)
}
\arguments{
\item{id}{The identifier code of a Eurostat data product.}
}
\value{
A list that contains three tables: the indicator, a value label
description table and a metadata table.
}
\description{
Get a Eurostat data product and save its metadata and the data in
tidy tables.
}
