% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_indicators.R
\name{impute_indicators}
\alias{impute_indicators}
\title{Impute Indicators}
\usage{
impute_indicators(indic)
}
\arguments{
\item{indic}{An imputed indicator from \code{\link{get_eurostat_indicator}}.}
}
\value{
A data frame that with updated actual, missing, approximated, forecasted, nocb, locf estimates.
}
\description{
In the following order tries to fill missing values:
}
\details{
\enumerate{
\item Approximate missing values withing time series with \code{\link{na_approx}}
\item Next observation carry forward for old missing values \code{\link{na_nocb}}
\item Forecast the time series ahead \code{\link{indicator_forecast}}
\item If the forecast did not work, try last observation carry forward \code{\link{na_locf}}.
}
}
\seealso{
Other approximation functions: 
\code{\link{na_approx}()},
\code{\link{na_locf}()},
\code{\link{na_nocb}()}
}
\concept{approximation functions}
