% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{indicator}
\alias{indicator}
\alias{is.indicator}
\alias{print.indicator}
\title{Indicator for an Automated Data Observatory}
\usage{
indicator(
  x,
  shortcode,
  indicator_code = NULL,
  indicator_name = NULL,
  description,
  description_at_source = NA_character_,
  date_indicator = NULL,
  date_earliest = NULL,
  date_latest = NULL,
  original_source = NA_character_,
  code_at_source = NULL,
  doi = NULL,
  keyword1,
  keyword2,
  keyword3,
  keyword4,
  keywords = NA_character_
)

is.indicator(x)

\method{print}{indicator}(x, ...)
}
\arguments{
\item{x}{A data.frame or tibble with at least \code{geo}, \code{time},
\code{value}, and \code{estimate} columns.}

\item{shortcode}{A short, unique, programatically usable, memorizable indicator
ID, given by an observatory curator.}

\item{indicator_code}{A machine-generated code. If there is no machine-generated code present,
it is filled with \code{shortcode}.}

\item{indicator_name}{A unique name for the indicator.}

\item{description}{A precise character string describing the indicator
for the data catalogue of the data observatory.}

\item{description_at_source}{A machine read description from the source, may require
manual revision.}

\item{date_indicator}{The creation/refreshment date of the indicator. If omitted, calls
\code{Sys.Date()}.}

\item{date_earliest}{The date of the earliest observation. Can be a numeric containing a year.}

\item{date_latest}{The date of the latest observation.}

\item{original_source}{Defaults to {NA_character_}.}

\item{code_at_source}{The identifier in the original source, if applicable.}

\item{keyword1}{The first keywords must be on of the observatories: \code{economy},
\code{greendeal}, \code{music} and planned ones: \code{competition}, \code{creative}.}

\item{keyword2}{The second keyword must be one of the pillars of the
observatory.}

\item{keyword3}{The third must be a topic within a pillar.}

\item{keyword4}{A search term within the topic.}

\item{keywords}{A character vector of any optional, further keywords.}
}
\value{
A data.frame or tibble with indicator attributes.
}
\description{
The class inherits all methods from a data frame, but has
many fixed attributes.
}
\examples{
test_indicator <- indicator (
                     x <- data.frame ( 
                     geo = rep(c("NL", "BE", "LU"), 4), 
                     time = rep(c(2016:2019),3), 
                     value = runif(12, 1,100), 
                     estimate = rep("actual", 12)
                     ), 
   shortcode = "observatory_test_1", 
   description = "A test indicator with random numbers", 
   date_indicator = as.Date ( "2020-08-24"),
   date_earliest  = min (x$time, na.rm=TRUE),
   date_latest  =  max(x$time, na.rm=TRUE),
   keyword1 = "music",  keyword2 = "economy",  keyword3 = "demand", keyword4="pcr"
)

## Only the first 10 observations are printed 
print (test_indicator)
}
