% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_database.R
\name{initialize_database}
\alias{initialize_database}
\title{Initialize a database}
\usage{
initialize_database(path = ":memory:")
}
\arguments{
\item{path}{A path}
}
\description{
Initialize a database
}
