% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_variables.R
\name{is_time}
\alias{is_time}
\title{Test time Column Formatting.}
\usage{
is_time(time)
}
\arguments{
\item{time}{An indicator table to test.}
}
\value{
\code{TRUE} if the test is met, otherwise \code{FALSE}.
}
\description{
The time dimension of an indicator must be a Date, or
it must be a year as an integer, or a numeric that
can be coerced into an integer.
}
\examples{
{
test_indicator <- data.frame (
   geo = c("DE", "DE", "CH", "CH"),
   value = 1:4,
   time = as.Date(paste0(2020:2021, "-01-01")),
   estimate = rep("actual", 4)
 )
is_unique_observations(test_indicator)
}
}
\keyword{internal}
