% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_unique_observations.R
\name{is_unique_observations}
\alias{is_unique_observations}
\title{Test Unique Observations}
\usage{
is_unique_observations(indicator)
}
\arguments{
\item{indicator}{An indicator table to test.}
}
\value{
\code{TRUE} if the test is met, otherwise \code{FALSE}.
}
\description{
Tidy indicators have observations that are unique. They must have on of the three
types of values: an actual, an estimated or a missing value.
}
\details{
Approximation and other filling techniques require unique observations.
}
\examples{
{
test_indicator <- data.frame (
   geo = c("DE", "DE", "CH", "CH"),
   value = 1:4,
   time = as.Date(paste0(2020:2021, "-01-01")),
   estimate = rep("actual", 4)
 )
is_unique_observations(test_indicator)
}
}
