% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jevons_aggregation.R
\name{Jevons_aggregation}
\alias{Jevons_aggregation}
\title{Jevons static aggregation}
\usage{
Jevons_aggregation(data)
}
\arguments{
\item{data}{dataframe with rows = observations and columns = quantitative
variables}
}
\value{
It returns a dataframe with rows = observations and column =
composite indicator
}
\description{
The Jevons_aggregation function computes an index using the Jevons method of
static aggregation. This method calculates the geometric mean, multiplied
for 100.
}
\examples{

data("Education")
Indicator=Jevons_aggregation(Education)
print(Indicator)

}
\references{
Massoli, P., Mazziotta, M., Pareto, A., Rinaldelli, C. (2013).
Synthesis Methodologies and Spatial Analysis. Composite indices for BES,DAYS
OF RESEARCH IN ISTAT, NOVEMBER 10-11, 2014
}
\keyword{Jevons}
