% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_aggregation.R
\name{linear_aggregation}
\alias{linear_aggregation}
\title{Linear Aggregation}
\usage{
linear_aggregation(data, weights = rep(1/ncol(data), ncol(data)))
}
\arguments{
\item{data}{dataframe with rows = observations and columns = quantitative
variables}

\item{weights}{vector of weights (default all weights = 1/ ncol(dataframe))}
}
\value{
It returns a dataframe with rows = observations and column =
composite indicator
}
\description{
This is a function to apply linear aggregation index
}
\examples{

data("Education")
Indicator=linear_aggregation(Education)
print(Indicator)

}
\references{
OECD/European Union/EC-JRC (2008), Handbook on Constructing
Composite Indicators: Methodology and User Guide, OECD Publishing, Paris,
<https://doi.org/10.1787/9789264043466-en>
}
\keyword{Aggregation}
\keyword{Linear}
