% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_indicator_eurostat.R
\name{map_indicator_eurostat}
\alias{map_indicator_eurostat}
\title{Map Eurostat indicators into a databse}
\usage{
map_indicator_eurostat(ids, path)
}
\arguments{
\item{ids}{Identifiers of Eurostat statistical products.}

\item{path}{A path to save the database.}
}
\description{
Map Eurostat indicators into a databse
}
\examples{
\dontrun{
map_indicator_eurostat <- function (
 ids = c("ISOC_R_BLT12_I", "isoc_cicce_use", "teicp090"),
 path = 'not_included/example.db' )
}
}
