% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate.R
\name{na_locf}
\alias{na_locf}
\title{Last observation carry forward}
\usage{
na_locf(indicator)
}
\arguments{
\item{indicator}{A tibble created by \code{\link{get_eurostat_indicator}}. with time, geo,
value, frequency, estimate and method columns.}
}
\value{
A tibble updated with the forward carried values.
}
\description{
Last observation carry forward
}
