% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_aggregation.R
\name{rank_aggregation}
\alias{rank_aggregation}
\title{Ranking Aggregation}
\usage{
rank_aggregation(data)
}
\arguments{
\item{data}{dataframe with rows = observations and columns = quantiative
variables}
}
\value{
It returns a dataframe with rows = observations and column =
composite indicator
}
\description{
It's a function to apply ranking aggregation. The highest value is the first
value of rank
}
\examples{

data("Education")
Indicator_rank=rank_aggregation(Education)
print(Indicator_rank)

}
\references{
OECD/European Union/EC-JRC (2008), Handbook on Constructing
Composite Indicators: Methodology and User Guide, OECD Publishing, Paris,
<https://doi.org/10.1787/9789264043466-en>
}
\keyword{Rank}
\keyword{aggregation}
