% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_normalisation.R
\name{rank_normalisation}
\alias{rank_normalisation}
\title{Rank normalization}
\usage{
rank_normalisation(data)
}
\arguments{
\item{data}{dataframe with rows = observations and columns = quantiative
variables}
}
\value{
It returns a datafame of normalized data
}
\description{
It's a function that normalize by ranking method. Create the matrix of
ranking for different columns, the rank is the high value is the first
}
\examples{

data("Education")
Normalized_rank=rank_normalisation(Education)
print(Normalized_rank)

}
\references{
OECD/European Union/EC-JRC (2008), Handbook on Constructing
Composite Indicators: Methodology and User Guide, OECD Publishing, Paris,
<https://doi.org/10.1787/9789264043466-en>
}
\keyword{Rank}
\keyword{normalization}
