% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_population.R
\docType{data}
\name{small_population}
\alias{small_population}
\title{Population of Small Countries}
\format{
A data frame with 60 rows and 5 variables:
\describe{
\item{unit}{NR, natural count}
\item{sex}{F for female, M for male, and T for total population.}
\item{geo}{Geographical observation area, countries}
\item{time}{time of observation}
\item{values}{number of people}
}
}
\source{
\url{https://appsso.eurostat.ec.europa.eu/nui/show.do?dataset=demo_pjan&lang=en}
}
\usage{
small_population
}
\description{
A dataset containing male and female population (all age groups) of three small
countries, Andorra, Liechtenstein and San Marino from 2010 to 2020.
}
\keyword{datasets}
