% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_population_description.R
\docType{data}
\name{small_population_description}
\alias{small_population_description}
\title{Labelling of the Small Countries' Population Indicator}
\format{
A data frame with 3 rows and 10 variables:
\describe{
\item{shortcode}{A curated short code, or the indicator_code, if not available}
\item{description}{A curated description text.}
\item{keyword_1}{Keyword for observatory, for example \code{"music"}}
\item{keyword_2}{Keyword for pillar, for example \code{"economy"}}
\item{keyword_3}{Top level division of the pillar.}
\item{keyword_4}{2nd level division of the pillar.}
\item{futher_keywords}{Optional further keywords as a concantenated list}
\item{shortcode}{A curated short code, or the indicator_code, if not available}
\item{description_at_source}{A machine-created description from source descriptive metadata}
\item{indicator_code}{A machine-created unqiue indicator code.}
\item{original_source}{The source, if applicable, or our observatory}
}
}
\source{
\url{https://appsso.eurostat.ec.europa.eu/nui/show.do?dataset=demo_pjan&lang=en}
}
\usage{
small_population_description
}
\description{
The labelling metadata of the indicator at source, containing the total population of three small
countries, Andorra, Liechtenstein and San Marino from 2010 to 2020.
}
\details{
Processed with \code{\link{get_eurostat_indicator}}.
}
\seealso{
Other data files: 
\code{\link{small_population_indicator}},
\code{\link{small_population_labelling}},
\code{\link{small_population_metadata}},
\code{\link{small_population}}
}
\concept{data files}
\keyword{datasets}
