% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_population_labelling.R
\docType{data}
\name{small_population_labelling}
\alias{small_population_labelling}
\title{Labelling of the Small Countries' Population Indicator}
\format{
A data frame with 6 rows and 5 variables:
\describe{
\item{shortcode}{A curated short code, or the indicator_code, if not available}
\item{indicator_code}{A machine-assigned unique identifier for the indicator}
\item{var_code}{The original coding of a variable at source}
\item{var_label}{The original labelling of the variable at source}
\item{var_name}{The name of the variable at source}
}
}
\source{
\url{https://appsso.eurostat.ec.europa.eu/nui/show.do?dataset=demo_pjan&lang=en}
}
\usage{
small_population_labelling
}
\description{
The labelling metadata of the indicator at source, containing the total population of three small
countries, Andorra, Liechtenstein and San Marino from 2010 to 2020.
}
\details{
Processed with \code{\link{get_eurostat_indicator}}.
}
\seealso{
Other data files: 
\code{\link{small_population_indicator}},
\code{\link{small_population_metadata}},
\code{\link{small_population}}
}
\concept{data files}
\keyword{datasets}
