% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardization.R
\name{standardization}
\alias{standardization}
\title{Standardization}
\usage{
standardization(data)
}
\arguments{
\item{data}{dataframe with rows = observations and columns = quantitative
variables}
}
\value{
It returns a dataframe of scaled data
}
\description{
It performs a standardization of data, i.e., centering and scaling, so that the
data is expressed in terms of standard deviation (i.e., mean = 0, SD = 1).
When applied to a statistical model, this function extracts the dataset,
standardizes it, and refits the model with this standardized version of the
dataset
}
\examples{

data("Education")
Standardization=standardization(Education)
print(Standardization)

}
\references{
OECD/European Union/EC-JRC (2008), Handbook on Constructing
Composite Indicators: Methodology and User Guide, OECD Publishing, Paris,
<https://doi.org/10.1787/9789264043466-en>
}
\keyword{standardization}
