% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardization_MPI.R
\name{standardization_MPI}
\alias{standardization_MPI}
\title{Standardization of data with Maziotta-Pareto index}
\usage{
standardization_MPI(data)
}
\arguments{
\item{data}{dataframe with rows = observations and columns = quantitative
variables}
}
\value{
It returns a dataframe of normalized data
}
\description{
This is a function that standardized the data with Maziotta-Pareto index
}
\examples{
data("Education")
Standardization_MPI=standardization_MPI(Education)
print(Standardization_MPI)

}
\references{
De Muro P., Mazziotta M., Pareto A. (2011), "Composite Indices
of Development and Poverty: An Application to MDGs", Social Indicators
Research, Volume 104, Number 1, pp. 1-18
}
\keyword{MPI}
\keyword{Standardization}
