% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate.R
\name{test_unique_observations}
\alias{test_unique_observations}
\title{Test Unique Observations}
\usage{
test_unique_observations(indicator)
}
\arguments{
\item{indicator}{An indicator table to test.}
}
\value{
\code{TRUE} if the test is met, otherwise and error message.
}
\description{
Tidy indicators have observations that are unique. They must have on of the three
types of values: an actual, an estimated or a missing value.
}
\details{
Approximation and other filling techniques require unique observations.
}
