% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_indicator.R
\name{tidy_indicator}
\alias{tidy_indicator}
\title{Tidy Raw Indicator}
\usage{
tidy_indicator(indic_raw, indicator_labels = NULL)
}
\arguments{
\item{indic_raw}{A raw Eurostat indicator}

\item{indicator_labels}{The codes and description of \code{indic} columns.}
}
\value{
A tibble with fixed form.
}
\description{
The indicator downloaded from Eurostat is brought to a tidy format.
}
\keyword{internal}
