% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_metadata.R
\name{update_metadata}
\alias{update_metadata}
\title{Get Tidy Indicator(s) From Eurostat}
\usage{
update_metadata(indic, metadata)
}
\arguments{
\item{indic}{An imputed indicator}

\item{metadata}{A metadata table returned by \code{\link{get_eurostat_indicator}}.}
}
\value{
A data frame that with updated actual, missing, approximated, forecasted, nocb, locf estimates.
}
\description{
Get a Eurostat data product and save its metadata and the data in
tidy tables.
}
\details{
This function creates a tidy indicator table that is ready to be inserted into a database.
}
\seealso{
Other acquisition functions: 
\code{\link{get_eurostat_indicator}()}
}
\concept{acquisition functions}
