% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate.R
\name{indicator_forecast}
\alias{indicator_forecast}
\title{Forecast the indicator value}
\usage{
indicator_forecast(indicator)
}
\arguments{
\item{indicator}{A tibble created by \code{\link{get_eurostat_indicator}}.}
}
\value{
A tibble updated with with forecasted values.
}
\description{
Forecast the indicator value
}
