% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_eurostat_metadata.R
\name{load_eurostat_metadata}
\alias{load_eurostat_metadata}
\title{Load Eurostat metadata}
\usage{
load_eurostat_metadata(envir = globalenv())
}
\arguments{
\item{envir}{The environment to load the metadata, defaults to the
global environment of your current R session.}
}
\description{
The function loads metadata dictionaries and the table of contents
into the global environment to avoid repeated requests to the Eurostat
data warehouse.
}
