% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate.R
\name{na_approx}
\alias{na_approx}
\title{Linear approximation of missing values}
\usage{
na_approx(indicator)
}
\arguments{
\item{ids}{Identifiers of Eurostat statistical products.}

\item{path}{A path to save the database.}
}
\value{
A tibble updated with with approximated values.
}
\description{
Linear approximation of missing values
}
