% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate.R
\name{na_approx}
\alias{na_approx}
\title{Linear approximation of missing values}
\usage{
na_approx(indicator)
}
\arguments{
\item{indicator}{A tibble created by \code{\link{get_eurostat_indicator}}. with time, geo,
value, frequency, estimate and method columns.}
}
\value{
A tibble updated with with approximated values.
}
\description{
Linear approximation of missing values
}
