% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate.R
\name{na_nocb}
\alias{na_nocb}
\title{Next observation carry back}
\usage{
na_nocb(indicator)
}
\arguments{
\item{indicator}{A tibble created by \code{get_eurostat_indicator}.}
}
\value{
A tibble updated with the values carried back.
}
\description{
Next observation carry back
}
