% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_unique_observations.R
\name{test_unique_observations}
\alias{test_unique_observations}
\title{Test Unique Observations}
\usage{
test_unique_observations(indicator, stop_on_error = FALSE)
}
\arguments{
\item{indicator}{An indicator table to test.}

\item{stop_on_error}{Defaults to \code{TRUE} when the code stops with an error
message.  If \code{FALSE}, it displays non-unique values.}
}
\value{
\code{TRUE} if the test is met, otherwise and error message.
}
\description{
Tidy indicators have observations that are unique. They must have on of the three
types of values: an actual, an estimated or a missing value.
}
\details{
Approximation and other filling techniques require unique observations.
}
