% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate.R
\name{test_unique_observations}
\alias{test_unique_observations}
\title{Test Unique Observations}
\usage{
test_unique_observations(indicator)
}
\arguments{
\item{indicator}{An indicator table to test.}
}
\value{
\code{TRUE} if the test is met, otherwise and error message.
}
\description{
Approximation and other filling techniques require unique observations
}
